+++ app/routes/circulars/circulars.server.ts
@@ -7,7 +7,11 @@
  */
 import { tables } from '@architect/functions'
 import type { DynamoDB } from '@aws-sdk/client-dynamodb'
-import { type DynamoDBDocument, paginateScan } from '@aws-sdk/lib-dynamodb'
+import {
+  type DynamoDBDocument,
+  paginateQuery,
+  paginateScan,
+} from '@aws-sdk/lib-dynamodb'
 import { search as getSearch } from '@nasa-gcn/architect-functions-search'
 import {
   DynamoDBAutoIncrement,
@@ -21,6 +25,7 @@ import { type User, getUser } from '../_auth/user.server'
 import {
   bodyIsValid,
   formatAuthor,
+  formatCircularText,
   formatIsValid,
   parseEventFromSubject,
   subjectIsValid,
@@ -31,7 +36,7 @@ import type {
   CircularChangeRequestKeys,
   CircularMetadata,
 } from './circulars.lib'
-import { send, sendEmail } from '~/lib/email.server'
+import { sendEmail, sendEmailBulk } from '~/lib/email.server'
 import { feature, origin } from '~/lib/env.server'
 import { closeZendeskTicket } from '~/lib/zendesk.server'
 
@@ -40,6 +45,8 @@ type Require<T, K extends keyof T> = Omit<T, K> & Required<Pick<T, K>>
 
 export const submitterGroup = 'gcn.nasa.gov/circular-submitter'
 export const moderatorGroup = 'gcn.nasa.gov/circular-moderator'
+
+const fromName = 'GCN Circulars'
 
 const getDynamoDBAutoIncrement = memoizee(
   async function () {
@@ -591,4 +598,63 @@ function validateCircular({
   if (!bodyIsValid(body)) throw new Response('body is invalid', { status: 400 })
   if (!(format === undefined || formatIsValid(format)))
     throw new Response('format is invalid', { status: 400 })
+}
+
+async function getEmails() {
+  const db = await tables()
+  const client = db._doc as unknown as DynamoDBDocument
+  const TableName = db.name('circulars_subscriptions')
+  const pages = paginateScan(
+    { client },
+    { AttributesToGet: ['email'], TableName }
+  )
+  const emails: string[] = []
+  for await (const page of pages) {
+    const newEmails = page.Items?.map(({ email }) => email)
+    if (newEmails) emails.push(...newEmails)
+  }
+  return emails
+}
+
+async function getLegacyEmails() {
+  const db = await tables()
+  const client = db._doc as unknown as DynamoDBDocument
+  const TableName = db.name('legacy_users')
+  const pages = paginateQuery(
+    { client },
+    {
+      IndexName: 'legacyReceivers',
+      KeyConditionExpression: 'receive = :receive',
+      ExpressionAttributeValues: {
+        ':receive': 1,
+      },
+      ProjectionExpression: 'email',
+      TableName,
+    }
+  )
+  const emails: string[] = []
+  for await (const page of pages) {
+    const newEmails = page.Items?.map(({ email }) => email)
+    if (newEmails) emails.push(...newEmails)
+  }
+  return emails
+}
+
+export async function send(circular: Circular) {
+  const [emails, legacyEmails] = await Promise.all([
+    getEmails(),
+    getLegacyEmails(),
+  ])
+  const to = [...emails, ...legacyEmails]
+  await sendEmailBulk({
+    fromName,
+    to,
+    subject: circular.subject,
+    body: `${formatCircularText(
+      circular
+    )}\n\n\nView this GCN Circular online at ${origin}/circulars/${
+      circular.circularId
+    }.`,
+    topic: 'circulars',
+  })
 }
